Rename providers to repository
