+++ scrapy_poet/utils.py
@@ -11,9 +11,7 @@ from scrapy.utils.defer import maybeDeferred_coro
 from twisted.internet.defer import inlineCallbacks, returnValue
 
 from web_poet.pages import ItemPage, is_injectable
-from scrapy_poet.providers import (
-    providers,
-)
+from scrapy_poet.repository import providers
 from scrapy_poet.page_input_providers import PageObjectInputProvider
 
 _CALLBACK_FOR_MARKER = '__scrapy_poet_callback'
