+++ tests/test_utils.py
@@ -4,9 +4,7 @@ from typing import Any, Dict
 import scrapy
 from scrapy.http import TextResponse
 
-from scrapy_poet.providers import (
-    provides,
-)
+from scrapy_poet.repository import provides
 from scrapy_poet.page_input_providers import (
     PageObjectInputProvider,
     ResponseDataProvider,
